Refactor journal forms and models
This commit refactors the journal forms and models by renaming the `create_entry_form` to `create_journal_entry` in the `__init__.py` file. It also adds a new file `create_journal_entry.py` that contains the `JournalEntryForm` class for creating a new journal entry. Additionally, the `thumbnail_directory_path` function in the `collection.py` file is updated to use the `uuid` instead of the `id`.